Make the remote-docs.sh script actually work
7458 was a good start but I'm afraid the developer never actually
tried to run the script and diff -urq against a prior run. The
script was failing catastrophically on darwin, because of my
(Ed's) suggestion based on a complete misunderstanding of how
the script is invoked.

Anyhow, this restores past functionality and, as hoped for,
removes the need for maintenance by using 'podman help' to
catalog subcommands. The key is that we need to look for .1
files for linux/mac, but .1.md files for windows. This
now handles both.

Also: improved fail() by including script name prefix, and
removing unnecessary usage message. Made script fail if
a particular doc file is not found.

Signed-off-by: Ed Santiago <santiago@redhat.com>